colorpicker: better integrate with canvasevent
Some code love for the color picker:

* Remove hardcoded modifier keys stuff now it can be launched with any
  combination. It still does a "modifiers still held" test, so record
  which modifiers were held when the window appears.

* Both styles of grab fix different bugs, so do both. Other code adopts
  this strategy too.

* Split up the code flow into different event handlers a bit. Slight
  voodoo here, but ensuring the window is shown fully before grabbing
  ay be more correct.

Partly ameliorates https://gna.org/bugs/?20358 by fixing the "first
launch does not update" problem with the stylus. The rogue grab for
second and subsequent stylus launches of the picker, which as far as I
can tell is device specific, is *not* fixed, but can at least now be
undone by drawing a stroke on the canvas.