Build: Don't attempt to link against _mypaintlib.so Python module
Instead link libmypaint-test into the Py module, and provide a
function for running the test that can be called from Python,
and a wrapper .py script that does so.
This also means we don't have to embedd our own Python interpreter
into the executable.
Removes a non-portable symlink hack from the build system.