canvasevent: workaround color picker grab release
Dispatch ColorPopupWindow as if it was launched from the menu if it was
actually invoked by [<Modifierss>+]ButtonN. This means you don't get
the popup window when picking unless you use the keyboard and hold the
key, but it avoids a fairly horrid _device specific_ grab release
problem that can't be addressed directly with PyGTK.

https://gna.org/bugs/?20358

Binding to a pointer button is still possible, but because we no longer
grab you can't drag around to adjust the colour finely, and nor do you
get a preview. You just have to keep clicking.

Of note; at least one user has expressed aesthetic concerns about the
picker's appearance, and that they'd prefer not to see the picker at all
if it's distracting.

http://forum.intilinux.com/mypaint-development-and-suggestions/annoying-colour-picker-behavior/

This workaround isn't intended to address that concern (sorry!) it's
just intended to create a system that doesn't lock users out of menus
and tool windows. Hopefully this will just be a temporary measure.