color picker: grab fix: destroy window after use
It seems the spurious and stylus-specific grab that was hanging around
after colour picking is cured by destroying the popup window after the
pick is done.

Refactor color picker to create its window for the duration of the pick,
destroying it afterwards. Reenable its normal call in the canvasevent
dispatcher.

https://gna.org/bugs/?20358

There are still some cases when stylus-specific grabs persist after the
picker is done, but the situation is now much more managable for users
aassuming they don't want to pick outside the drawWindow or in the
palette window or something. I suspect these are now in the domain of
actual GTK (or WM) bugs, and hopefully fixable when we migrate to GTK3.