move-layer: fix a race condition with leave()
Leaving due to a springloaded mode keyup was causing leave() to be
called while idle callbacks were still processing the finalize code,
leading to an exception. Fix by forcing everything to completion during
leave() while the mode is still in a good state.

https://gna.org/bugs/?20397

Also spotted and fixed a similar one with the move idler.