brush parser exception handling
Print the name of the brush that could not be loaded.

Catch the BrushInfo.ParseError exception, because it will usually
happen before the GUI exception handler is active, with the result
that the program doesn't start any more. (Ideally, we would show
a dialog at least, and give the user some hints how to remove the
brush(es) in question.)