gtk3: fix picker getting stuck in grab
GI+GTK3 reports button numbers strangely when peeking at the last event
on the stack (something horrid we need to do when responding to a
non-keyboard-iniaiated Action, e.g. a mouse menu selection or a toolbar
button press). Record a null button press initially, and let the first
subsequent button-press record the number so that the release works.

This code is getting somewhat hairy. Do we need the preview thing at
all? It's cute, but I count two other colour previews on my screen most
of the time. It's also somewhat slow for GTK3 mode, perhaps because the
screen pixel reading code itself is slower? Definitely breaks the
feeling of continuity in a hold-and-drag pick.

For fullscreen mode, when it matters, could it be converted to an
optional fading overlay that just triggers and then fades out whenever
the colour changes? Perhaps near the pointer, perhaps in a corner.
Something similar for brushes might be nice too.