cursors: hotspot fix; reduce min size; even sizes
All in the name of making the new cursors feel nice and precise.

https://gna.org/bugs/?20415

Add back even sized cursors, to be more like the old bitmap+mask set.
Not certain why this was off; perhaps to reduce the number of cases
while discussing feel and precision (odd sizes are simpler).

Change the default colours back to pure black and pure white.

Experimentally reduce the minimum size to 3px for more precision (one
white pixel surrounded by a ring of black). We should make this
configurable via the prefs dialog IMO because it's *way* too small for
proper usability. But some expert users seem to like flyspeck cursors :D

Move the hotspot 1px up and left. This should be the bullseye of an
odd-sized cursor, and is probably a bugfix for that.
For an even-sized cursor, it's the North-West pixel of of the 4
surrounding the exact centre.

I think, if the feel is right for odd-sized cursors now, we could adjust
the freehand painting position to the South-East by half a pixel for
even sized cursors.