brush popup: moving cursor out closes
Close the brush popup dialog (with a reject code) when the user moves the
pointer away from it. Use several pixels of slack around the edges to allow
the dialog to be resized under certain WMs and DEs, and to allow for
less-than-perfect moue coordination.

Bugfix: closing the window no longer destroys the dialog.

Minor stuff: add docs and comments.