freehand cursors: add presets, add crosshair style
Add prefs UI on the View tab for cursor presets, and set thin, medium, and
thick styles of circle. For the circle styles, "light outline" is the
default, a white circle inside a black one with some alpha.  It suffers
from a hardware rendering bug on some systems, so add "medium" and "heavy"
outline options with black borders to all sides at a couple of extra
weights.  Good compromises, hopefully.

Add a simple crosshair style for those who just want to know where the nib
of their pen is. It uses icons to show the fancier brush blend modes.