mask editor: update New dialog, allow luma changes
Allow the Luma of the HCY mask editor to be changed, to make it more useful
when considering colour schemes fr dark and light values. Scroll wheel
will do for now - do we need a Reference Luma slider?

Luma in these sub-dialogs is based on the working brush's luma, but
constrained between 0.25 and 0.75. Same thing for mask load-save dialogs.

Update the New mask dialog to make mask colours which suit the current hue
distortion (RYB, RYGB, ...)