preview window: get ready for integration
Fixups and code tidying for alown's <ali@lown.me.uk> excellent preview
window addition to work with the post-1.1.0 changes. Thanks, Ali!

Some functional and code changes here:

- Rotation bug fixed.
- Hook into the new view_changed_observers interface for updates.
- Overlay minimises screen updates.
- Widget zooming shows all of the main view rectangle, as well as the
  document bbox.
- Click to move the main-view rectangle because it's simpler, though
  drag works too.

This UI isn't completely settled yet, and perhaps we need an arrow to show
the "up" direction on the overlay, and any mirroring (suggestions, please!)
But it's working well enough to be integrated now, hopefully :)