tdw: remove redundant observer interfaces
Remove the defunct rotation and movement observer interfaces from
gui.tileddrawwidget. No problem with spying on changes to individual TDW's
transformations, but it should be done as a single observer interface and
not broadcast details of individual rotations and translations. It's better
to keep the state in one place.

For now, there's view_changed_observers on the main gui.document.Document,
which does the same thing. It could be moved to the TDW if needed; it's
only in doc as an evolutionary accident.

Reverts changes introduced in fddaf9dc6dc042b44bf28411fb78ea417589aee7