prefs: move dialog contents to GtkBuilder XML
Define the bulk of the prefs UI in GtkBuilder, editable in glade3 (hence
the different extension). It was getting unwieldy, and had broken
completely with the migration to GTK3.

This is pretty much GTK3-only code now. GtkGrids, and combo boxes with
string ID columns probably aren't backwards compatible with PyGTK.

Support argless construction for ButtonMappingEditor.

Add translation support for strings in gui/*.glade files.