tdw: make buildable without args
Make the view (gui.tileddrawwidget) buildable without specifying a model
(lib.document) at construction time. Instead, feed in the model to use with
set_model() after construction. If it's not set, draw a blank screen.

Move responsibility for creating the model and the view from the
controller (gui.document.Document) to the code which creates a
model/view/controller cluster.