Add weakref-friendly syntax sugar for observers
Python decorators for defining observer list interfaces in a way that's a
bit more public and formal and documentable than a plain old list member.
Uses some cute C#-like syntax for tacking on callbacks and additional
behaviour as tersely as possible.

Observable "@event"s set up this way store weakrefs to their observer
objects if the callback is a bound method, to allow them to die off
naturally.  Not needed right now, but it'll be necessary later on to allow
dynamic (parameterized, and shorter-lived) tool widgets to observe the
brushmanager, for example.