workspace: parameterizable building from layouts
Allow tool widget "tabs" in the workspace to be constructed with static
parameters. These parameters can't be changed over the lifetime of the
widget, and get saved and restored in the usual way.

This is intended to allow the brush and colour selectors to be split up
across multiple tabs: parameters will hold the brush group name, or the
colour selector subtype.