Tabbed workspace support for brush groups
Let each brush group have its own tab in the workspace. The icon used is
that of the first brush in the group, but that can probably be improved.

Support brush group creation, import, export, rename, delete from various
menus and popup dialogs.  All the stuff that the soon to be deleted
GroupSelector currently does, in other words.