brusheditor: tidy up brush settings editor GUI
New look for the brush settings editor, reusing all of its code into a new
module. The editor now takes the form of an index tree with the settings on
the left, a main area with base values and curve foldouts, and lots of
lovely whitespace and headings and indent. It should be more comprehensible
to new users like this, hopefully!

For my Xfce4 desktop at least, subwindows in fullscreen need to EITHER
reject keyboard input entirely and be added to the keyboard manager, OR
accept it and be fully modal dialogs with properly set transient-for hints.
Mixing the two styles leads to Great Confusion in fullscreen mode.  This
applies to things like spinners even, so the brush input curves now use
scalebuttons. Hopefully OK there...

The GUI layout for this window is now defined in a GtkBuilder XML file,
editable in glade (see glade/run.sh). It's a bit of an evolving hack given
that we have to template it for all the fiddly little per-input widgets,
but it seems to work sanely.