Fix brush settings not updating until editor shown
Fix regression: key presses to control radius, opacity etc., and the
dropdown controls for the same things were not working until the brush
editor window was shown for the first time. Fixed by moving the callback
binding for these GtkAdjustments to the main application instance.

Document the dict that contains these things, since that's the primary
interface for accessing them.