load_ora(): enable frame if it was customized
Enable or disable the frame when loading an OpenRaster file. The frame is
enabled if it's not the same as the calculated MyPaint image size, i.e.
the extent of the image's tiles. Avoids having to save a flag in stack.xml,
at the expense of some possible user confusion.

Users who change the frame extents hopefully get what they want when an
image is loaded now. Users who want an infinite sketchpad and no frames on
loading should also be happy.

To revert from a customized frame to a default one, set it to the bounding
box of the document. There's a button for that in the dialog, but perhaps
it needs a "reset everything" button too?