external layers: add load from/save to ORA
Add a basic OpenRaster load and save subsystem for non-bitmap data which
GdkPixbuf knows how to support. SVG is pretty much the only useful example
of this sort of data right now.

The vague idea here is that past revisions of the external file to be
edited will reside in the layer's private tempdir for the duration of a
session, and that editing will be modal so that MyPaint's undo and redo
work.  However, this isn't set in stone just yet: could add a file watcher
interface instead, at a later date.