canvasevent: turn off stroke queueing
Stroke data queueing (and its cheesy interpolation) wasn't liked by our
testers, so turn it off by default. We can remove it later on if the GTK
team provide a fix for motion compression, or if we come up with a proper
solution ourselves; conversely, we can make it an option if that's not
going to happen.

https://mail.gna.org/public/mypaint-discuss/2013-09/msg00000.html

Event queueing is still enabled, since it speeds things up measurably and
doesn't introduce any slow-tracking-like behaviour.

Switch to collections.deque rather than list for queues, again in the name
of reducing overhead during event processing. Deques are just under twice
as fast as lists for queue and stack behaviour.