fill: feedback fixes: cursor setting & sensitivity
The options widget shouldn't be made insensitive. Bad idea: the user may
want to adjust stuff even if filling won't work right now due to a layer
being locked or the cursor being outside the frame.

Fix the cursor being incorrect when the mode is entered by a key press.