frame: use proper overlay, show editable elements
Move drawing the frame to an overlay object, and make it aware of the mode
stack. If the user is currently editing the frame, draw the adjustable
edges and corners and highlight them if the pointer is near them.

Some minor adjustments to the hit-boxing tests to work better with a
visible frame.

Rationale: makes the operation of FrameEditMode more obvious.

Side effects: frame will no longer be drawn on the preview window. Is this
likely to be an issue? More redraws and calculations within FrameEditMode.