workspace: make paned contents prettier
Add visible frames implemented as GtkScrolledWindow wrappers for the
sidebar and canvas Paned slots to improve the distinctiveness of these
elements from their flanking toolbar and statusbar.

We now automatically repack the central canvas's widget hierarchy to make
its frame vanish when entering fullscreen mode, and undo the change when
leaving.  Users will want the full extent of the screen when fullscreened.

Using a GtkScrolledWindow and not a GtkFrame for the canvas makes the UI
more consistent (my test themes assign different appearances to frames and
scrolledwindows inside paned). It also raises the possibility of "merely"
exposing a couple of GtkAdjustments on the TiledDrawWidget to allow real
scrolling, at least when MyPaint is not in fullscreen mode.  A somewhat
dynamic system where the visible extents are the actual document bbox plus
one sensibly-sized page might work acceptably (i.e. patches welcome).