input events: add event filter to improve capture
Add an experimental GDK event filter to capture events before GDK itself
can motion-compress them and thus break our drawing accuracy. See

https://gna.org/bugs/?21003
https://gna.org/bugs/?20822
https://bugzilla.gnome.org/show_bug.cgi?id=702392

Good coding practice dictates that we don't try to do too much inside this
fairly horrid little hack, so we only queue the pointer position and the
timestamp. For all the events GDK 3.8+ drops on the floor during a frame
clock cycle (which the filter now captures), we use the tilt and pressure
info from the event on the event that did make it through. Interpolating
would be better.

Before: 50-60 events/second for all devices.
After: 200 events/second with an Intuos 5. 80 with a touchpad.

This hack only works on X11 platforms right now