eventhack: acquire GIL, bump data tuple refcount
Ubuntu 13.10 (and maybe other systems) require the GIL to be held during
then event filter, so wrap the filter in the appropriate calls. Not
required for all systems, but it's good practice to do this given that the
filter uses Python API calls outside the scope of the Python interpreter,
where execution may or may not be threaded. Fixes a segfault on affected
systems.

http://docs.python.org/2/c-api/init.html#non-python-created-threads

Also bump the data tuple's refcount for the duration of the filter's
operation, if the display type is supported. Technically not needed given
the framework in which it's called, but will be safer in case of error.