eventhack: interpolate pressure and tilt axes
The eventhack code doesn't record pressure or tilt since that requires
knowledge of the maxima and minima for their raw readings in order to be
compatible with GDK's translated values. Such code would be fragile, so
instead interpolate their values based on the detailed eventhack times and
the time and axis values from the coarse events that make it through GDK's
motion compression.

Move pressure and tilt faking and sanity checking up front, since we now
interpolate using the sanitized values from this phase.