style: EXPERIMENTAL configurable toolbar icon size
Symbolic icons are designed for 16x16, but we'd like to use them on the
toolbar too, for which 24x24 is optimal. Users also like to see the 16x16
size instead, and cram more onto the bar. Allow the size to be configured
via a prefs setting (no UI for it right now).