inputtestwindow: improve reporting
For some reasons I now get motion events reported from both tdw and
drawWindow with gtk+ from git. Only count tdw events when reporting
the motion event rate. Seems to be only an issue in the test window.

Add more context so it's possible to see what is going on.