quick choosers: add color popup, common base
The dialog-based popup chooser for the brush chooser was proving annoying
with my window manager, so switch to a resizable popup-window based
implementation which uses the same kind of pointer leave mechanism for
cancelling.

Add a colour chooser popup based on the same thing, for some visual
similarity. That goes on "C"; remove the old color changer on "V"
temporarily. The colour chooser is also accessible by clicking on the
bottom left "current colour" button.