model: working-doc tempdir, and cleanup at exit
Add a working-doc tempdir which persists for as long as a document is open,
i.e. from opening one document to opening the next. This tempdir
persists for as long as the undo stack (and its actions) have a meaningful
relationship with the loaded document.

External layers are expected to keep loaded files at various undo points in
the temp dir. It may be useful to add detection and cleanup of unreachable
disk state to things that drop off the undo and redo stacks too, in order
to prevent this area growing too big. This will do for now though.

Automatically clean up this area at application exit, and when
transitioning from one doc to the next. Both are guarded by dialogs.