layer: implement copy(), use for Duplicate Layer
Add a generic but customizable copy() method for layers, for use by
DuplicateLayer. Mostly just shorthand for save/load snapshot, but we also
replicate layer flags.

Background layers cannot currently be copy()ed.

Implement save/load snapshot for ExternalLayer so that it can be copied.
The copy and snapshot semantics both demand an independent copy of the
file, so make load_snapshot() perform a copy within the document working
directory.