layerswindow: use a tree-based view of the stack
Use a GtkTreeView rather than a GtkListView for the layer stacking tree.
Update the associated display-side cache objects accordingly.

Beginnings of drag-and-drop for reparenting.

Various method renames.

It's only possible to move a single layer at a time with any UI, so lose
the lazy "assign this new order" command in the backend.

Use an icon to denote sub-stacks.

Implement compositing of sub-stacks.

Moving a layer "into" another layer may result in both layers being
children of a newly created parent. All undoable.