layers window: various look and feel tweaks
Eye and lock icon to the right hand side. Type icon or expander in a column
to the left.

Allow layers to invoke a type-specific action when their type icon is
clicked on. The LayerStack class now expands or collapses itself as its
action so that its expander icon can occupy the same column as other
layers' type action icon.

Sizing and ellipsis tweaks.

Add expander icons.