duplicate layer: extend support to sub-stacks
Allow sub-stacks to be cloned with Duplicate Layer.

Simplify the undo stack command for Duplicate Layer to work with layer
paths only, and use only the current layer.

Add an interface for giving a layer a unique name, and begin using it when
duplicating.

Semantics have changed: snapshots must now save and restore everything the
user would expect in Duplicate Layer, including the name.