prefs: UI updates; large toolbar icons by default
Space out the prefs window a bit more, and add an explanatory note to the
bottom of the View page telling users they'll have to restart to see the
icon size and dark theme variants.

The "large" (24x24, same size as before) icons are prettier, so turn them
on by default. There is a known problem with Ubuntu 12.04 LTS (and similar
versions of GTK) where larger symbolic icons do not render correctly. This
can be worked around by upgrading or by selecting small icons.