app: use a less glaringly white bg by default
Remove the default shipped "default.png" for now, and pick a default bg
colour based on how bright the UI is assumed to be, but not so dark that if
we get it wrong the bg looks strange.

I dunno. Maybe we should use an actual paper texture out of the box? Maybe
something subtly off-grey and polychrome so that users can experiment with
colorize etc.