OpenRaster conformance: nobble some group features
To fully conform to the current OpenRaster draft specification, rendering
flags like opacity or compositeop need to be disallowed on layer stacks,
which implies that layer groups must not render in isolated mode, to use
terminology from the W3C Compositing-1 specification.

For now, since I'd like to get the other parts of this crazy refactor back
into master, we must silently prevent the user from changing, loading,
saving, or seeing certain layer properties if the layer is a layer group.
In other words, turn off the new, useful things that layer groups would
allow until the specification can be updated, a process which will require
agreement from the GIMP and Krita devs.