Generalize old "stroke splitting" call
* Refactor stroke splitting as a general input flush request
* Refactored freehand painting accordingly
* Refactored geometric lines tool accordingly
* Refactored layer moving accordingly
* Fixes https://gna.org/bugs/?21854

Switch model "stroke splitting" mechanism over to a more general
replacement covering all types of undoable change to the working document.
What used to be done as a "stroke split" should now be done as a flush
request on the model, flush_updates(), which pulls in changes from anywhere
registered to provide them. This naturally includes the current interaction
mode, and may include other sources of document updates.

Doing this allows highly interactive tools which don't draw strokes (e.g.
Move Layer, Fill) to finalize what they're doing when the user invokes a
command which does something else to the document (e.g. Clear Layer).

The neatest way of using this new mechanism this is to retool the undo
stack commands for interactive modes to do work between their construction,
and their committing onto the CommandStack. The document is then updated
interactively via the "live" but uncommitted command object so that it can
keep track of everything that happens.