observable: tidy up repr() forms
Make repr()s which contain details of bound methods not require the
underlying object of the method to be repr()-able yet. This is because
observers can be set up during instantiation of the underlying object, and
it's unduly strict to require them to be fully repr()able at this point.

Also, brevity is nicer.