layers: remove old layer expanded/collapsed state
The expansion state of a layer lived in the model, and this was dubious.
Whether the layer is expanded or collapsed in the GUI is a matter for the
GUI components alone. It's OK for the model itself to signal that it would
like the interface to expand or collapse a path, however, and it would be
OK to reflect (but not store) expansion state in the model one day should
we choose to save it.

Remove expand and collapse management options from layer bubbling. This
command should only have one codepath and it should do the Right Thing™.

(This commit will break things a bit until the new layerswindow code in the
pipeline starts using the new expand/collapse events)