Update unique layer naming
If the user tries to rename a layer with a null name, they get the default
name for this kind of layer as a starting value.

When the UI displays a layer in the layers group with a null name, use the
default name there too. There's something better than this around the
corner; for now it's just safest to make the view code not permute the
model.

Layer names are uniquified when added to the root stack using deepinsert().

Remove ancient name='' args during layer handling in various places.