root stack: can now observe current layer property
Make changes to the current layer an observable event in the lifecycle of
the root layer stack.

Also make the get/set property functions more robust in the face of bad
data. In particular, the property can now be set to None to reset it -
although you'll only see None as a value from the getter if the document is
empty, which should never be the case.

deeppop() now tries to preserve the current layer path.
deepremove() does the same, and now returns None (like it should).