layer stacks: reverse storage order
Storing layers in rendering order is becoming increasingly bad for any
purpose other than rendering, so reverse it.

Layers are now stored in order from the highest layer in the stack to the
lowest. That's the same as OpenRaster's reading and writing order, and the
same as GTK's tree layout order in a GtkTreeView.

- ORA loading and saving is fixed accordingly in this commit.

- The necessary GUI rework is deferred to a later commit :) having all tree
  paths use the same raw lists of indices will be a HUGE win for finer
  grained layers window updates and general GUI code quality.

- Raise/lower/up/down commands, pick order, backdrop rendering for Sample
  Merged fill all inverted.

- Tree structure notifications from the model now follow the assumed GUI
  rendering order.

This commit also gets rid of some dubious postorder traversals of the layer
stack in certain places.