layer: fix args to canonpath, add doctest
RootLayerStack.canonpath()'s position-based fallback for non-empty stacks
is now "usefirst", i.e. fall back to using the highest layer in the stack.
This follows the inverting of the layer path addressing, but arguably this
is better for the user than the old fallback.

Update all uses of the old fallback ("uselowest").

Add unit test for canonpath().