Code robustness: drag and selection
Fix a rare race-condition exception during treeview updates which can
(sometimes) be triggered by repeated reloading interleaved with bashing
away at other actions. Basically do not assume that the layer data is
available in the treeview row data functions.

If the selected layer is completely removed during a drag, updates which
query it can raise an exception if they're not carefully coded. Fix three
related exceptions which can be revealed by attempting to drag the last row
of a flat structure such that it becomes a nested structure.