Brushwork command: fix a potential exception
Fix exception when clearing a document a lot while also drawing to it in
freehand mode with event queues. This only replicably happens in the brush
preview editor (for slightly mysterious reasons), and it doesn't matter
about screwing up the undo stack there - so just log the problem for now.