drawwindow: switch filename and app name order
* Update the format string for titlebars with filenames
* Also mark the strings for translation, just in case

I've noticed that most of the programs on my desktop suffix filenames with
the app name, not vice versa.  This has the effect of putting the more
user-relevant information, namely the file basename, at the front of the
titlebar string.  That allows crowded window lists in panels to display
more of the user-relevant information despite their truncation of titlebar
strings.