Add a GtkAccelMap editor
The ability to rebind the global accelmap directly from the menus is gone
permanently in GTK 3.12+, which prefers that you migrate to GtkApplication
and manage your accelerators through its API.  However due to API
mismatches, binding editors working via GtkApplication cannot be
implemented simply AND completely AND in a way that's compatible with 3.10.

We'd like to target 3.10 while not having accels break in 3.12+.  Something
has to change.  Sorry, Martin ☹

For now, add an editor for the legacy global AccelMap to the dialog so that
accels can be edited under both GTK+ 3.10 and 3.12+. The UI we'll need for
this after migrating to GtkApplication will be pretty similar to this old
GtkAction stuff, so it's hopefully not a wasted effort.

https://gna.org/bugs/?21751