tests: add raw rendering performance test
Per #32, the scroll performance tests are a bit useless now. Part of what
that was good for was testing how fast rendering a layer stack is.

Add a new test that does an off-screen emulation of scrolling in a circle
with a big document on a 1080p screen to address #21. It also exercises the
Cairo rendering path. Should be good for testing future tile caches, and
improvements to the C++ tile compositing code.