display rendering: use a LRU cache
Rendering for the display involves a fix15 → 8bpp transformation after
generating each fix15 tile. Tile generation is a moderately expensive
operation compared to conversion, so cache that fix15 data in RAM using a
cache with an LRU metric.

The ideal maximum capacity for the cache remains to be determined, but 32
or 64 MiB has a nice feel, and shouldn't cause modern systems any trouble.

Addresses mypaint/mypaint#21.