Simplified the functionality of eraser mode
It now no longer changes size by itself. This is more of the nuclear
option than anything, as a lot of brushes DO need to have their size
changed to work well as an eraser, but the previous code was so
user-hostile and counter-intuitive that it was best to entirely remove
it.