Support fake, hopefully faster, checkerboard bgs
Optimizing the part of the rendering path involving rendering RGBA data
with A<1 is proving intractable. "RGBU" (i.e. RGBA with A=1) is noticeably
faster. Therefore use the tile rendering phase to fake a checkerboard
background.

New setting: "Prefs → View → Zoom → Use a real checkerboard pattern".
The faked kind is a little ugly, and is prone to dithering artefacts. Also
its appearance changes with the mipmap level, which looks odd. Therefore
it's off by default.

Addresses #21 by trading a little raw performance for hopefully much less
lag when a fake transparency checkerboard background is turned on. CPU use
for 1:1 zoom + a single destination-in layer on my system drops from 160%
to 120%.