Improve debugging some more
Looks like we'll need the source as well to distinguish some multi-faceted
devices, i.e. tablets sporting touchpads, pad buttons, or eraser pens in
addition to pen input, all of which which present as separate devices.

Quite a lot of non-Wacom tablets present all of these devices under the
same name when used with the evdev driver.  See the output people have
reported in #29.

Getting the references we store right will be important for preferences
driven inclusion or exclusion of devices, bug #40.