Fix build when using clang on Linux
The build environment was assuming gcc on the "linux2" platform, or at
least it was using "-z origin" when the linker was being invoked via clang
wrappers. Replaced with passthru variant for better portability.

We should be testing this to help OSX users and porters (see #49). The
current way of compiling with a nonstandard compiler is via the standard
environment vars CC and CXX:

   $ CC=clang CXX=clang++ scons

Builds with clang 3.4 and gcc 4.9.0 under Debian testing/unstable.