Refined Two Symbolic Icons and Patched Export Script
Refined the mypaint-tab-options, and the mypaint-fill symbolic icons to
help the wrench flow with the theme, and the paint bucket look a bit
better(it the small things).

Also, it was getting anoying that one had to close Inkscape in order to
finish the symbolic icon export. So adding a --verb=FileQuit command did
the trick to close inkscape when it was finished.

Lastly, notice that for some of the symbolic icons, the export script
did not remove the bounding box of the icon. Fixed by adding a loop
to looked for something that was common between the bounding boxes
which they all start with "use". Simple Band-Aid fix, but I plan to open
and issue about an alternate method on how to export the symbolic icons.