Run lib/meta.py to obtain release version
Grepping  a special line out via the shell and eval'ing it is a dirty way
of getting the base version string, and gets in the way of PEP8 compliance.
Instead, run the metainfo module as a python script. Currently this just
prints the value of MYPAINT_VERSION to STDOUT, but we could use the eval
trick again if we need more info to be available to the release-time
scripting.

Fixes #113.

(Also remove .travis.yml from the tarballs we'll distribute.)